requirements.txt
streamlit
plotly
pandas
streamlit-shadcn-ui
databricks-sql-connector
